Created user registration and login frontend service.
To access, login to server then /signin. Users should be able to
register and sign in to account. Currently working without a backend,
storing users in local storage (go to web inspector and look at DOM and
cookies for testing). Backend service for this will be next.
BUG: After a user logs in, the user should be redirected to user page
(dashboard for example), but there is some weird behavior that doesn't
cause correct redirect. However, if you examine cookies you can see that
the login response is successful.