Connected frontend user service to backend (Round 1)
Now you can go to sign up or sign in in landing page and register an
account. Disabled CSRF configurations for now since I was having an
issue transporting the CSRF token between sessions. After registering,
you can see user is created in db. I also had to modify primary key for
users in db to be auto incremented. This was causing me many problems
when inserting users to db, but not anymore. Need to fix redirections
and create a login authentication method in backend next.