Added/Deleted some Help Page and Index Page
I added some information to the Help page, it is static right now but
probably doesn't need to be dynamic.
Also deleted the signup and login from the landing (index) page. So it
just has our title and about information. also changed the css for it
too match our upcoming changes to dashboard.